+++ moabb/datasets/huebner_llp.py
@@ -43,6 +43,8 @@ class _BaseVisualMatrixSpellerDataset(BaseDataset, ABC):
         use_blocks_as_sessions=True,
         subjects=None,
         sessions=None,
+        *,
+        return_all_modalities=False,
         **kwargs,
     ):
         self.n_channels = 31  # all channels except 5 times x_* CH and EOGvu
@@ -56,6 +58,7 @@ class _BaseVisualMatrixSpellerDataset(BaseDataset, ABC):
             subjects=(np.arange(n_subjects) + 1).tolist(),
             selected_subjects=subjects,
             selected_sessions=sessions,
+            return_all_modalities=return_all_modalities,
             **kwargs,
         )
 
@@ -384,6 +387,8 @@ class Huebner2017(_BaseVisualMatrixSpellerDataset):
         use_blocks_as_sessions=True,
         subjects=None,
         sessions=None,
+        *,
+        return_all_modalities=False,
     ):
         llp_speller_paper_doi = "10.1371/journal.pone.0175856"
         super().__init__(
@@ -397,6 +402,7 @@ class Huebner2017(_BaseVisualMatrixSpellerDataset):
             use_blocks_as_sessions=use_blocks_as_sessions,
             subjects=subjects,
             sessions=sessions,
+            return_all_modalities=return_all_modalities,
         )
 
 
@@ -629,6 +635,8 @@ class Huebner2018(_BaseVisualMatrixSpellerDataset):
         use_blocks_as_sessions=True,
         subjects=None,
         sessions=None,
+        *,
+        return_all_modalities=False,
     ):
         mix_speller_paper_doi = "10.1109/MCI.2018.2807039"
         super().__init__(
@@ -642,6 +650,7 @@ class Huebner2018(_BaseVisualMatrixSpellerDataset):
             use_blocks_as_sessions=use_blocks_as_sessions,
             subjects=subjects,
             sessions=sessions,
+            return_all_modalities=return_all_modalities,
         )